requirements.txt
Django>=3.2,<4.0
django-ckeditor>=6.0,<7.0
googletrans==4.0.0-rc1
redis>=4.0,<5.0
django-redis>=5.0,<6.0
djangorestframework>=3.12,<4.0
